microservices/app/app/src/cleanraj/requirements.txt
appdirs==1.4.3
certifi==2016.9.26
dj-static==0.0.6
Django==1.9.5
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-user-accounts==1.3.1
elasticsearch==1.9.0
gunicorn==19.6.0
image==1.5.3
jsonfield==1.0.3
olefile==0.44
packaging==16.8
Pillow==4.1.0
pinax-eventlog==1.1.1
pinax-theme-bootstrap==7.8.0
pinax-webanalytics==2.0.1
pyparsing==2.2.0
pytz==2016.4
simplejson==3.10.0
six==1.10.0
static3==0.7.0
urllib3==1.19.1
requests==2.18.4